move data struct from api to store
