--- api/memo_relation.go
@@ -1,15 +1,0 @@
-package api
-
-type MemoRelationType string
-
-const (
-	MemoRelationReference  MemoRelationType = "REFERENCE"
-	MemoRelationAdditional MemoRelationType = "ADDITIONAL"
-)
-
-type MemoRelationCreate struct {
-	// Standard fields
-	MemoID         int              `json:"memoId"`
-	RelationMemoID int              `json:"RelationMemoId"`
-	Type           MemoRelationType `json:"type"`
-}
