+++ server/memo_relation.go
@@ -7,7 +7,6 @@ import (
 	"strconv"
 
 	"github.com/labstack/echo/v4"
-	"github.com/usememos/memos/api"
 	"github.com/usememos/memos/store"
 )
 
@@ -15,19 +14,19 @@ func (s *Server) registerMemoRelationRoutes(g *echo.Group) {
 	g.POST("/memo/relation", func(c echo.Context) error {
 		ctx := c.Request().Context()
 
-		memoRelationCreate := &api.MemoRelationCreate{}
+		memoRelationCreate := &store.MemoRelationCreate{}
 		if err := json.NewDecoder(c.Request().Body).Decode(memoRelationCreate); err != nil {
 			return echo.NewHTTPError(http.StatusBadRequest, "Malformatted post memo relation request").SetInternal(err)
 		}
 
-		if memoRelationCreate.Type != api.MemoRelationReference && memoRelationCreate.Type != api.MemoRelationAdditional {
+		if memoRelationCreate.Type != store.MemoRelationReference && memoRelationCreate.Type != store.MemoRelationAdditional {
 			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid Relation Type: %s", memoRelationCreate.Type))
 		}
 
 		message, err := s.Store.UpsertMemoRelation(ctx, &store.MemoRelationMessage{
 			MemoID:        memoRelationCreate.MemoID,
 			RelatedMemoID: memoRelationCreate.RelationMemoID,
-			Type:          api.MemoRelationType(memoRelationCreate.Type),
+			Type:          store.MemoRelationType(memoRelationCreate.Type),
 		})
 		if err != nil {
 			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Create Relation Fail: %v %v", memoRelationCreate.MemoID, memoRelationCreate.RelationMemoID)).SetInternal(err)
@@ -52,7 +51,7 @@ func (s *Server) registerMemoRelationRoutes(g *echo.Group) {
 	g.DELETE("/memo/relation/:memoId/:relatedMemoId/:type", func(c echo.Context) error {
 		ctx := c.Request().Context()
 
-		relationType := api.MemoRelationType(c.Param("type"))
+		relationType := store.MemoRelationType(c.Param("type"))
 		relationDelete := &store.DeleteMemoRelationMessage{
 			Type: &relationType,
 		}
@@ -68,7 +67,7 @@ func (s *Server) registerMemoRelationRoutes(g *echo.Group) {
 			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Relation memo ID is not a number: %s", c.Param("memoId"))).SetInternal(err)
 		}
 
-		if relationType != api.MemoRelationReference && relationType != api.MemoRelationAdditional {
+		if relationType != store.MemoRelationReference && relationType != store.MemoRelationAdditional {
 			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid Relation Type: %s", relationType))
 		}
